move ReceiverInitError to where it is used
